[flow] Kill `React$AbstractComponent` for real
Summary: Changelog: [errors] `React.ComponentType<Props>` is now only an alias of `component(...Props)`, instead of some special cased types. This comes with stricter checks and conversions, such as making Props readonly, erroring on the presence the ref prop instead of silently ignoring them, and ensures that `Props` is a subtype of `{...}`. In addition, the `React$AbstractComponent` type is removed.

Reviewed By: panagosg7

Differential Revision: D71197053

fbshipit-source-id: 3cbb2a9c09f3977efe626ef4a52ccc9af6412821