[flow][ref-as-prop] Make React functions over components be generic over the presence/absence of ref prop
Summary:
This diff changes the signature of `React.lazy` and `React.memo` in order to make it generic over the presense/absense of ref prop. It's necessary to make the ref prop less special-cased in Flow.

Changelog: [lib] `React.lazy` and `React.memo` is now generic over the presence or absence of ref prop.

Reviewed By: panagosg7

Differential Revision: D71742536

fbshipit-source-id: 3356f44125bc13d9d664b29f96b6071e4dc01a7d