[flow] add "includes" flowconfig option for full mode of natural_inference_local_primitive_literals
Summary:
Rolling out Full mode of natural inference for primitive literals will incur a large number of errors. To make it easier to rollout, we are adding an option
```
experimental.natural_inference.local_primitive_literals.full.includes=<PROJECT_ROOT>/dir
```
that enables Full mode for files under `dir/`.

This flag does not require `experimental.natural_inference.local_primitive_literals` to be provided.

It also implies that files that are not in `dir/` will be checked in Partial mode.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71785414

fbshipit-source-id: e0093437d7108cd201a450a9506a10fc601feb52