[flow] Fix stale libdef-error issue
Summary:
Because we union together error set in types_js from the libdef init phase, if we don't have an entry in the error map for a file that previously has an error, then the old error will stick around. To fix this, we need to ensure to generate a `FilenameMap` with all files in the libdefs, so that it can override the previous error entry.

Changelog: [fix] Fix a bug that causes fixed `libdef-override` error to stick around after an incremental recheck.

Reviewed By: panagosg7

Differential Revision: D72079297

fbshipit-source-id: c403e88d3cb6ccd6c9444275cb644067e46d83b2