[flow] Always take the last definition of `declare class` as identity of classes in libdefs
Summary:
To resolve the issue in the previous diff, we make sure that the same class in scoped and common libdefs have the same identity, by always taking the last definition as the source of nominal identity.

Longer term, once we have declaration merging for declare classes, we no longer need this hack, because we can see it operating in a principled way: earlier definitions are all augmenting the last definition.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D72083729

fbshipit-source-id: a877976d02f76e70fe7bbe3a288dc57d1a911488