[flow] More specific error code for react-rule-hook 1/n
Summary:
In this stack of diff, I aim to provide specific error code for different kinds of react-rule-hook errors, so that it's easier to audit different kinds of violation, and maybe even making some of them unsuppressible in the future.

Changelog: [errors] For react-rule-hook errors related to naming convention issues, it will now have `react-rule-hook-naming-convention` error code.

Reviewed By: panagosg7

Differential Revision: D72203431

fbshipit-source-id: 12192b3b26bbfe05bf2ddb58c595e72b4e7f9516