[flow] More specific error code for react-rule-hook 2/n
Summary:
In this stack of diff, I aim to provide specific error code for different kinds of react-rule-hook errors, so that it's easier to audit different kinds of violation, and maybe even making some of them unsuppressible in the future.

Changelog: [errors] For react-rule-hook errors related to conditional hook calls, it will now have `react-rule-hook-conditional` error code.

Reviewed By: panagosg7

Differential Revision: D72260244

fbshipit-source-id: af639b094455cf559ced36a128571351c208e91e