[flow] generalize_singletons descend into Promise<>
Summary:
In code like
```
class AsyncSelector<T>  {
  constructor(getState: () => T) {}
}
const x = new AsyncSelector(async () => ({title: 'blah'}));
```
we'd like the type for `'blah'` to be inferred as `string`, which means that we need to generalize that portion in the inferred type.

To do this, `generalize_singletons` needs to descend into the type that corresponds to the inferred `Promise<...>`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D72324488

fbshipit-source-id: f1dc8b7aabe277aee464643a6e25c130f3abf2c2