[flow] keep precise literal types when checked
Summary:
In natural inference full mode, in implicit instantiation calls like the following
```
function foo<V: 1|2>(x: V): V;
foo(1);
```
we need to be able to determine that parameter `1` needs to be inferred in its precise form, so that it passes the check against the bound of type parameter `V`.

With the existing support for generalization, we do not have enough information to make this decision.

This diff fixes this by adding a mechanism that tracks what checks are done over string (and other primitive) literals.

In terms of implementation these checks can happen at various points, but are mostly concentrated around:
* `Subtyping_kit`
* `Flow_js_utils.ground_subtype`
* `Speculation_kit.optimize_spec_try_shortcut`

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D71363665

fbshipit-source-id: f95328b55446a7375ebded968141bdad8949ec52