[flow] add support for const type parameters
Summary:
This diff adds support for const type parameters.

Given the definition
```
declare function fn<const X>(x: X): X;
```
the call `fn({a: 1, b: ['b']})` will return the type `{+a:1, +b: $ReadOnly<['b']>}`.

In other words the result will be equivalent (*) to having passed in `{a: 1, b: ['b']} as const` as argument to the call.

The asterisk (*) above is due to fact that we've kept the behavior that the reasons on the various literal containers (objects, arrays) that are converted to const form are still going to be "literal" reasons. This is a mostly for compatibility with existing code. If we had stripped the literal reasons then we'd lose compatibility with mutable versions of these types which are often used as annotations. E.g.
```
const obj: {a: 1, ['b']} = fn({a: 1, ['b']});
```
The more sound behavior (that strips the literal reasons) can be enabled with the `~keep_container_reasons` parameter of `convert_literal_type_to_const`.

## Sanity checks

Const type parameters are only allowed on functions, methods and classes, and are incompatible through subtyping with non-const type params. This can change later on if it becomes impractical.

## Implementation

The conversion to const form happens right after implicit instantiation has pinned the type for a const-parameter `X`. If `X` is declared as `const`, we call into `convert_literal_type_to_const`, a mapper that extends the existing `literal_type_mapper` used for generalization, but with the appropriate `singleton_action`.

The mapper in `convert_literal_type_to_const` handles specifically the cases of object and array literals which need to become readonly.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision:
D72762802

------------------------------------------------------------------------
(from 47806afdf97585a19634797816a6fc638e693dd1)

fbshipit-source-id: 743505aabd68867c12a5d38eedafeb5de1219540