[flow] fix unary expression checking in full natural inference
Summary:
When inferring a type for the expression `-42`, we recursively call `expression` on `42`. The flags we pass to this call need to reflect how the result is being used. Specifically, if there is a hint for `-42` then this information needs to be passed along to the inner expression. Similarly if this is an assignment to a const-variable.

This diff fixes this by passing these flags appropriately. Note that for the hint to work as expected we need to install it in Name_def first.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D72828177

fbshipit-source-id: b903fec77c430e50c92ca135f13e66ab8d122285