[flow] convert_literal_type_to_const never keeps container reasons
Summary:
Before this diff in
```
declare function fn<const X>(x: X): X;
const obj = fn({a: 1});
```
the resulting type was `{+a:1}`, but the outer object had a "literal" reason which allowed unsound coercions to the type `{a:number}`.

This diff fixes this by making the reason of the outer object at const type parameter calls a non-literal one.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73157623

fbshipit-source-id: 8a2b6a7e77914f8a609acbffe3dd5573750bc4df