[flow] extend TypeAppT ~> TypeAppT bypass mechanism
Summary:
Extends the approach of D45816760 to descend into conditional types and `$ReadOnly`.

Context:

For performance reasons comparisons of types `Foo<A>` and `Foo<B>`, where `Foo` is a type alias, reduce to comparisons between `A` and `B`. Unfortunately, this is done without recording the fact that in `Foo`, `A` may appear under a `$ReadOnly`. So we were missing the opportunity to reduce this check to comparing `$ReadOnly<A>` and `$ReadOnly<B>`, which would have not raised an error in the above example.

D45816760 special cased certain forms of the body of `Foo`, mostly around mapped-types, to bypass the optimization. This diff extends this to some additional forms, commonly found in type definitions like the ones in stylex.

It also protects the recursion from termination issues using "seen" sets for OpenT and EvalT ids.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73125867

fbshipit-source-id: a48be66b2976b78eab973e15e86165c8d6ad0456