[flow][refactor][EZ] Rename `declaring_function_component` to `in_context_possibly_expecting_fn_component_or_hook` and document it
Summary:
This stack tries to cleanup the logic in react_rules to make it easier to follow, so that I can confidently make some changes that tighten the rules.

The new name more accurately describe what is actually does (notably that it also assumes that we can be in hooks).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D73216421

fbshipit-source-id: b83982951723ac69a9ddf0dae04e5dd246f4d470