[flow][multiplatform] Replace string with Flow_import_specifier for resolved_require information 1/2: everywhere but shared memory
Summary:
See D73146584 for motivation for `Flow_import_specifier`. This diff makes it replace the `string` type that's used to represent these import specifiers with `Flow_import_specifier.t`.

It's done almost everywhere except for shared-memory, so that the shared-memory changes can be easily reverted if it's causing issues.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D73146585

fbshipit-source-id: 00aa74cecf80152d2984f0f0fe6d3e3f3dffb2c9