Restrict ppxlib version to < 0.36.0 (#9277)
Summary:
This PR restricts the ppxlib dependency to < 0.36.0 because the current code does not compile with ppxlib 0.36.0.

I'm not sure if the Flow team prefers this version constraint or would rather update the code to support the new ppxlib API. I'm happy to help with this if that is the preferred approach. Let me know what you'd like to do.

Pull Request resolved: https://github.com/facebook/flow/pull/9277

Reviewed By: SamChou19815

Differential Revision: D73367427

fbshipit-source-id: 0bf1d35259b0c1c392e1aa8b11d459b5da66eaa1