[flow] fix: do not reset primitive_literal_checks map
Summary:
The `primitive_literal_checks` map was added in D71363665 to keep track of checks done over string (and other primitive) literals during implicit instantiation to decide if we need to keep them in their precise form.

That diff would reset this mapping after every instantiation to empty. This had as a result checks made during synthesis (for example) in a call, to be cleared out, before the actual checking of the same call. In the event that some checking constraint was already cached from synthesis, then the checks that involved the primitive literals types would never be restored.

This diff fixes this by not reseting this map. This will be cleared once at the end of the lifetime of the current file context. This mapping only get updated during implicit instantiation so it should remain relatively small.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73415228

fbshipit-source-id: 33d271a1a40698212af3bd840a4a002637d6b37d