[flow] enable const-type-params by default
Summary:
Now that hermes-parser supports const type parameters, we can enable them in general.

Changelog: [feature] Added support for `const` type parameters that allow you to specify that an argument to a function is always treated as a const-expression, by annotating the type parameter with the `const` modifier. This eliminates the need to use the `as const` syntax when calling functions with this annotation. See more in [our docs](https://flow.org/en/docs/types/const-expression/#const-type-parameters).

Reviewed By: gkz

Differential Revision: D73470012

fbshipit-source-id: a06d557e88091720da698adf3efdb32a52c49bd2