[flow] add better version of Primitive_literal.convert_literal_type for nested generic calls
Summary:
This diff together with the following two diffs aim to address some edge cases in nested generic calls (or JSX constructions) under full natural inference.

Consider the code
```
declare function f<T: string>(v: T): C<T>;
declare function g<T: string>(v: C<T>): C<C<T>>;

const x1 = g(f('a')); // should be C<C<string>>
const x2: C<C<'a'>> = g(f('a'));
```
Today the second call does not typecheck under full natural inference. There are a few reasons that lead to this behavior. Some of these are expected, while others are suspect, and will be address in this diff and the following one.

1/ The call to `f('a')` does not impose any constraints on the type inferred for `'a'` -- treating it as a general `string` suffices, given the bound of `T`. This behavior is correct and will not be changing.

2/ There is no hint provided at the return value of the call to `f()` when post-processing the implicit instantiation call. (In reality, a hint does exist, but as of this diff there is no mechanism informing `generalize_singletons` of it.) Therefore, `Primitive_literal.convert_literal_type` will be called in `generalize_singletons` regardless of whether the additional precision will be needed for a check later on.

3/ If we took care of #2 and somehow propagated the information that the result of this call is passed directly to another call, and therefore potentially is checked against an annotation, the process taking place under `convert_literal_type` is problematic: it doesn't account correctly for structure of the resulting type of the inner function. Remember that `convert_literal_type` was initially created with the expectation that it only be called with a "literal" type (ie a type inferred directly for a literal argument to a call).

We address these issues, starting with #3 in this diff. Given that #2 still remains an open problem, fixing #3 won't have an effect on unit tests for now, but it still is necessary for the overall solution.

This diff creates a new version of `convert_literal_type`, called `implicit_instantiation_literal_mapper`, which targets specifically the result of implicit instantiation. The idea here is that the return type of the generic call may include a subterm in which a type parameter has been instantiated with a literal type. We therefore use a traversal similar to the one used for normal substitution, but instead of replacing `GenericT`s, we call into `convert_literal_type` when we have detected a "literal" type.

This diff also introduces `is_literal_type` as a predicate for types inferred for literal expressions.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73416514

fbshipit-source-id: 1473d98b133825be85b6ab6b0a5b74c2219dba1a