[flow] Move `ObjectAssignFromT` out of constraint system
Summary:
Put a concretizer at the same point in flow_js to ensure we keep the same behavior.

I take the decision to leave out some of the react deep-readonly stuff, so that `Object.assign` is not longer special cased. This is consistent with how we treat react deeponly when they are flow to other types.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D73323823

fbshipit-source-id: e4acef23164334659467f966e0af12bf065b7afe