[flow] pass syntactic hint info to help generalization in generic calls
Summary:
This diff addresses #2 of the list of items of D73416514, namely informs `generalize_singletons` when a hint exists, so that it preserves precise types.

This diff threads that information in implicit_instantiation.ml. We avoid using `Type_env.has_hint`, as this function always returns `false` in non-checking mode, whereas we want to keep the behavior consistent among typing modes.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73237643

fbshipit-source-id: c5404e15244f09ae43fc0d2f070f3ab28f53d841