[flow][refactor] Localize the hacky fix cache for `Object.assign` checking and document why it's bad
Summary:
I have finally understood what this cache (originally added in D6433731) is doing and it's very broken. At the point it was first introduced, it might be working as intended, but speculation mechanism has changed so much, so that now its usage here is a complete mess.

I documented the likely intention in the code, and localize it so that the chaos will be contained within `Object.assign`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision:
D73537213

------------------------------------------------------------------------
(from 5583bbc5ff26d9934fecb945b6684d2d4b178362)

fbshipit-source-id: 7e80e418fa13c06eeb064ae37a2c6d8e3e56aeae