[flow] always record object expression hints
Summary:
Increase the presence of hints around object literals, so that they can be picked by the natural inference generalization mechanism.

For example, in
```
const o: {fn:() => 'a'} = {
  fn() {
    return 'a';
  }
};
```
we want to attach a hint on the object literal to ensure that we have enough information to decide if `'a'` needs to be inferred in precise form.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73829729

fbshipit-source-id: c70b2b02a621a6eb4bac5354cdf9e1c2dff5abcf