[flow] Distinguish 3 kinds of hooks not in component/hook error in error code and message
Summary:
Based on the previous work, we can finally start distinguishing two kinds of calling hook not in component/hook errors.

It's currently split into two different kinds of severity: one where we decide that there is no way it's called inside a hook/component, or the case where the hook compatibility mode is off. The latter one is less likely to be an issue, because it might indicate that component/hook migration is not complete yet.

However, the current implementation doesn't guarantee that the former isn't actually "definitely-not-in-component-or-hook", so this diff also adds in additional checks to rule out the possibility. Notably, we will only error on hook calls in those named functions that

1. don't follow hook name convention
2. don't follow component name convention

This diff gives these 3 cases different error messages and error code.

Changelog: [errors] When Flow decides that the hook call definitely doesn't happen within component or hook body, it will emit errors with code `react-rule-hook-definitely-not-in-component-or-hook`.

Reviewed By: panagosg7

Differential Revision: D73216424

fbshipit-source-id: 250152b96abb41af977ca72f41ab423ee5e9732f