[flow] treat React children array as literal
Summary:
Code like `<><fbt/>{'a'}</>` triggers a "ReactConfigCheck" between the inferred JSX props and the component config props. In the case of JSX fragments, however, it seems that the component config is based on the types that are passed as children. So, effectively, this check is a check between the inferred children Props and themselves.

When there's multiple children these are added in a synthetic array and so this check ends up being an invariant check.

This has worked fine so far since these types are identical. Under natural inference, the type of the jsx prop `'a'` ends up being the singleton type `'a'`, whereas the contribution to the component config ends up being generalized to `string`. The invariant array check led to an incompatibility and a spurious error:
```
... string [1] is incompatible with string [1] in array element of property `children`.
```
I'm not inclined to reconsider the rules for generalization in generic calls just to handle this case more gracefully, given that the orthogonal decision to use inferred argument types as the component's config is probably suspect here.

Instead, I'm making `RReactChildren` a "literal" reason to force covariant only subtyping. This is a safe decision. The children array is a synthetic array that shouldn't leak the element creation site.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D73886892

fbshipit-source-id: b4622a10502d5af8757b4bb6815336548d97e39b