[flow] Use contextual typing to have less react-rule-hook errors
Summary:
We can use contextual typing to decide that a given function is actually used as a component, so it can be treated as component for the purpose of react-rule-hook checking. This is very helpful in the following cases:

```
React.memo(() => {
  useState(); // previously error, now ok, because the lambda is contextually typed to be component
})
```

We can potentially do the same for hooks, like

```
declare function testHook(h: hook () => {}): void;
testHook(() => {
  useState();
})
```

However, that will lead to `react-rule-hook-incompatible` error, so I don't think we should just do it unless we also use contextual typing to eliminate the `react-rule-hook-incompatible`  error, so I leave it out.

Changelog: [internal] hook compatibility = off is not announced.

Reviewed By: panagosg7

Differential Revision: D73932392

fbshipit-source-id: d876c001d541c9084058530921c4c6159fd9c643