[flow] Unwrap assertion operator on LHS (other than in typechecking)
Summary:
The `!` operator can occur on the LHS of assignments, which means that we now have another valid case for `Ast.Pattern.Expression`, unfortunately. Fortunately in most cases we can just pass through the operator and behave as if it was the argument of the node. This diff adds an unwrapping operator to all uses of assignment LHSes that I could find other than those in statement.ml, which will be handled separately.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D74055426

fbshipit-source-id: f0c14876b74148c8716b399ab357c0a9f3dbc023