[flow] improve hint on functions with rest params
Summary:
In hint decomposition for the return type of a function we create a constraint between the incoming function type `F` an a synthetic type that includes the target return type `R`:
```
F <: (...args: any) => R
```
The problem with this approach is that when `F` is of the form `(...args: []) => F_return`, then this comparison reduces due to `ResolveSpreadT(ResolveSpreadsToMultiflowSubtypeFull)` to
```
ArrT (ArrayAT _) <: ArrT (Tuple [])
```
which raises an `ENonLitArrayToTuple` error.

This error is unnecessary, since the only reason we introduced `...args: Array<any>` in the synthetic type was to avoid unrelated errors to hint decomposition. So, to avoid it we can just match the `rest_type` part of the type we use as input. This has better results in examples like the ones added as tests.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74343382

fbshipit-source-id: d9b59affa8da8ae6adaedb0c62c7c308784c0716