[flow] more consistent generalization of inferred type argument
Summary:
Under natural inference we expect in
```
declare function useWithBound<T: boolean>(x: T): [T, (y: T) => void];
const [o, set] = useWithBound(true);
```
`T` to be pined to `boolean` instead of `true`

Before this diff, the slight variation
```
declare function useWithBound_<T: void|boolean>(x: T): [T, (y: T) => void];
const [o, set] = useWithBound_(true);
```
would cause `T` to be pinned to `true` instead.

This was due to the shortcut check for `(l, UnionT (_, rep))` when `l` is in`quick_subtype` relation with a type in `rep` to always record this as a precise check on `l` and thus as a reason to keep the precise type for `l`.

This diff fixes this by only recording the check if we indeed have a singleton-to-singleton check succeeding under `quick_subtype`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74367855

fbshipit-source-id: 6907e76d6502331fc3a9f37857b878f76f802b71