[flow][react-rule] Use type information to find upper case non-component functions that call hooks
Summary:
This diff builds on top of the type aware analysis, to reject an upper case function with component-like name but doesn't return `React.Node` to be components. Thus, all the hooks call in such functions will have `react-rule-hook-definitely-not-in-component-or-hook` errors.

Changelog: [errors] When component syntax support is enabled, upper case functions with component-like name but doesn't return `React.Node` will no longer be treated like components. Thus, all the hooks call in such functions will have `react-rule-hook-definitely-not-in-component-or-hook` errors.

Reviewed By: jbrown215

Differential Revision: D74355595

fbshipit-source-id: abd2a999ccd2d1504c790683f183e41a8b4ca738