[flow] Typecheck optional chain nonnull operator
Summary:
Implement the remaining typechecking cases for `!`, for when it appears in an optional chain.

`!` in an optional chain works in a kind of funky way--it obeys the existing rules of optional chains in that `a?.b!.c` is not the same as `(a?.b)!.c`. If `a` and `a.b` are both nullable, `a?.b!.c` has a nullable type from `a`'s nullability but not `b`, while `(a?.b)!.c` is not nullable.

Fortunately this is easy to integrate into existing optional chain support. We add a new chain kind, `AssertChain`, which works similarly to `NewChain`, except that for optional chains, we pass a tvar to OptionalChainT which "collects" the nullish types in order to include them in the overall type of the chain, whereas `AssertChain` simply passes MixedT instead of the collecting tvar. This means that nullish types are noop flowed to Mixed and don't occur later anywhere in the type.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74076675

fbshipit-source-id: 38d8a8653ef7cd70889ce8cfc89f9569de6cd2e4