[flow] Typecheck nonnull operator in non-optional-chain positions
Summary:
This diff implements basic typechecking support for the ! operator in all positions other than as part of an optional chain (meaning a case where it appears within a truly optional context, so `a?.b!.c` is unimplemented but `a!.b!.c` is implemented).

For the most part this is simple--when we see `expr!`, we filter the type of `expr` to remove null and void.

It's slightly more annoying when `expr!` appears on the LHS. To support this, we need to drop the `!` operator from the LHS and then branch on whether the result is a member or an identifier, and then follow the same logic as the appropriate existing case. The structure of Ast.Pattern makes this more complicated than it might be, because unwrapping `x! = 42` gives us `Ast.Pattern.Expression (Ast.Expression.Identifier x)`, which is a structure that can't exist normally (instead we would have `Ast.Pattern.Identifier x`. So we end up duplicating the `Identifier` case logic to handle the two ways that identifiers can now appear on the left.

We also modify name_defs a little bit so that OpAssigns correctly populate the environment with the refined LHS. (Incidentally, operator assignment is the only place where ! on the left is meaningful, as far as I understand it: `x! = 42` seems silly but `x! += 42` makes sense if `x: ?number`)

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D74076674

fbshipit-source-id: 5239532d9e2520a671bb1c458f1c032a1944b635