[flow] use enclosing context for generalization in generic calls
Summary:
In code like
```
declare var obj: {a: number, b: string};
declare function getKey<T>(key: T): T;
obj[getKey('a')];
```
under natural inference there was no indication that the call was in a context where precise types are required.

This diff fixes this by storing the enclosing context of each call in a map so that implicit instantiation can access it, and force inference of precise types.

Arguably, there are more elegant ways to pass this information, but will leave as cleanup (T223795104).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74440178

fbshipit-source-id: 4e536ea42eedac56bbbcaec03c082c55bfb7b611