[flow] use has_hint instead of enclosing-context for jsx attributes
Summary:
When inferring types for jsx attributes we want to infer precise types for primitive literals. The implicit instantiation mechanism will revert those to general types later if needed to prevent leaking.

Before this diff we relied on enclosing context, with the `JsxAttrOrChildrenContext` variant, to pass that information when checking each attribute. The problem with this was that it was not applied deeply and so only the toplevel expression would benefit from it.
```
<Foo foo={'a'} />
```
For deeper levels, e.g.
```
<Foo foo={{bar:['a']}} />
```
where `foo: {bar:Array<'a'>}`, we had to rely on the regular type hint mechanism.

This mechanism works fairly reliably, but not for jsx-pragma, where hints are missing completely, likely because they were not very necessary for contextual typing purposes.

To make sure that we infer the right precision in the latest example under jsx-pragma, we set the more robust `has_hint` flag. This flags propagates more reliably in literal structures, and so handles more cases like the one above.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74453429

fbshipit-source-id: eeeb7d73512f73ff9efa2af3ef17c579993c1739