[flow] Config option for only allowing assert operator to be applied to arrays and indexer objects
Summary:
The general purpose assert operator is maybe too big and unsound of a hammer to allow general purpose use of everywhere. This limits the use of the operator to just array and indexer object lookups. This effectively means that `!` is a noop right now, unless it's an array with nullable type.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74207681

fbshipit-source-id: fd8d207d499d903709b76cd2c553926c8b483a97