[flow] estree: emit optional chain assertions as assertion expressions and regular optional chains
Summary: The estree format for optional chain assert operators that was being previously emitted was goofy and more importantly, incompatible with the Hermes parser (essentially, we emitted an OptionalMemberExpression with "assert" as the value for `optional`, while still emitting `true` or `false` in other cases). This changes the emitted code to always represent `!` with a `NonNullExpression`, and `NonNullExpressions` have a flag indicating whether they're optional chains or not. This aligns with how hermes parsing will work.

Reviewed By: SamChou19815

Differential Revision: D74214647

fbshipit-source-id: 0c909f4a657d2a5c563b954d07547e413e473506