[flow][multiplatform] Handle the .js.flow in common code case 2/2: interface conformance check
Summary: In this diff, we finished the final piece of Haste namespace support: common interface validation: when we have `Foo.js.flow` in the common code and `Foo.js.flow` or `Foo.js` in the web/native only code, we want to ensure that `exports of web/native only code ~> exports of common code`. This is similar to the common interface conformance check between `.ios.js`/`.android.js` and `.js.flow` for react-native multiplatform check.

Reviewed By: panagosg7

Differential Revision: D74751031

fbshipit-source-id: a1b7eb14ade6d09a75cad6253171688f04a35820