[flow][get-def] Track def loc for `module.exports`
Summary:
This diff starts to track def_loc of `module.exports` with best effort in type sig. In case we don't have a good syntactic one, we will fallback to the location in the reason.

Changelog: [fix] Go-to-definition on default import of `module.exports` will correctly jump to the exporting file

Reviewed By: panagosg7

Differential Revision: D74833422

fbshipit-source-id: 3e87db753184dfb340df729060c3e9a5299f5012