[flow][parser] Fix parser crash on starting hashbang
Summary:
Currently, on the given program in the issue, the peeked start location of the interpreter is `line 2, column -1`, which caused the crash when translating it into js token array.

This is a result of incorrect peek location for `T_INTERPRETER` when we add it into the token sink. `T_INTERPRETER` is parsed using comment infra which is not very careful with maintaining `Peek.loc`. In this diff, I just patch it by using the correct location attached to `T_INTERPRETER`.

Close https://github.com/facebook/flow/issues/9024

Changelog: [parser] Fix crash on `''#!/usr/bin/env node\n''` when generating token list

Reviewed By: panagosg7

Differential Revision: D74886013

fbshipit-source-id: 7fce58eab4f701b90fa33b1f385faa9269beaf1b