[flow][react-rule] Check that component props has subtype of `interface {}`
Summary:
This diff builds on the foundation of type-aware analysis of context of hook calls. Specifically, this diff refines the check that for the upper-case function that contains hook call, it's actually a react component. Previously, we check that the function returns `React.Node`. In this diff, we will also check that the function's first parameter is a subtype of `interface {}`.

Changelog: [errors] When component syntax is enabled, hook calls happen inside an upper case function that doesn't have a `props` param that's a subtype of `interface {}` will get `react-rule-hook-definitely-not-in-component-or-hook` error.

Reviewed By: panagosg7

Differential Revision: D74969063

fbshipit-source-id: 40d216b2a1262122582c4d37e48721882cc69cb3