[flow][match] fix: special enclosing context for Match pattern
Summary:
We want to infer types in the pattern part of a match expression/statement with most precision to enable optimal refinement results.

This diff fixes this by introducing a new kind of enclosing context to denote the context of a pattern in a match expression. The semantics of this context just instruct primitive_literal.ml to infer types in a precise form.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D75156191

fbshipit-source-id: 8bb79d01e2e197067112e41820119ce64d5434b4