[flow] relax failing hint decomp on empty rest array
Summary:
This tweak fixes a case of failing hint decomposition due to aggressively raising `UnconstrainedTvarException` when encountering OpenTs with "no lowers".

This was the case for the OpenT used as the rest array element type, even when the rest array has 0 elements. In that case, the OpenT is very likely to not accumulate any bounds. At the same time, this shouldn't affect the result of hint decomposition, since we're probably not going to be using this type.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D75585199

fbshipit-source-id: 65c6777ba0e14e4961c9b8a2d0da2ef246b7f369