[flow] consider RRestArrayLit due to calls as "literal" reason
Summary:
The goal of this diff is to mitigate the effect of a known issue with natural inference in generic calls. I have some thoughts on a more complete solution for the underlying problem, but for now this change helps unblock porting all unit tests to natural inference, and at the same time I think it is a reasonable approach consistent with existing practices.

Here's an mini repro of the issue with natural inference today
```
declare function foo<A: $ReadOnlyArray<mixed>>(...args: A): void;
foo(42);
```
This code returns a spurious error
```
Cannot call `foo` because  number [1] is incompatible with  number [1] in array element.
```
Here's what causes this:
1. We introduce a tvar `A'` and solve for it during implicit instantiation.
2. This yields a solution `ArrT([42])`, which thanks to generalization becomes `ArrT([number])`.
3. As part of the post-instantiation check, we create a synthetic type to represent the arguments passed to the call, `ArrT([42])`. Note that this type has not gone through generalization like the solution to the generic call. This was done on purpose. The type `42` and any other type inferred from primitive values will be in a "literal" position that enjoy co-variant checking even against writable structures (due to the implied uniqueness of the structure).
4. Where this fails in this example is that the array synthesized for arguments is not considered a "literal" array. Instead they are regular mutable arrays that cause invariant subtyping, and thus the invalid constraint `number <: 42`.

This diff tries to prevent this, by relaxing this final check: it makes the subtyping of the rest parameter array a "literal" one-way check. The correctness argument for this holds since there are no aliases for this synthetic array.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D74622179

fbshipit-source-id: 7cd60b27b7589a286984f8c18c2698906e6d6765