[flow][autoimport] Index local name of default import instead of inferring it from module
Summary:
In the previous diff, we have identified that we are not indexing the import name of default import properly. More specifically, we do not index the local name but instead uses inferred name of the module. Under that scheme, even if all the usages overwhelmingly use a name for default import, it will favor the import name inferred from the module name.

This diff changes that by using the local name instead. We just need to propagate the information already recorded in file_sig.

Changelog: [fix] For default imports, the autoimport ranking will now consider the names of the importing side. (e.g. Previously we completely ignore the name of `foo` in `import foo from './bar'`, but now we will count foo. If the pattern of `import foo from './bar'` happens a lot, then the autoimport algorithm will be more likely to suggest `import foo from './bar'` rather than `import bar from './bar'`).

Reviewed By: SamChou19815

Differential Revision: D75891244

fbshipit-source-id: 833895acef34d1aa090abc12138215130839c01d