[flow] purge BoolT_UNSOUND in Type_filter.not_truthy
Summary:
Eliminating uses of `BoolT_UNSOUND` in favor of `SingletonBoolT`.

This case is responsible for code like
```
const x = boolExpr ?? {};
```
where the result will be typed as `false|{}`, where `false` is non-generalizable. The reason we opt for the precise type for `false` is to account for the special-casing of "false" values as React props.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D75900510

fbshipit-source-id: 3fc7652290ee638de454aba0c3f5253e3baceb77