[flow] Create a single EPropsExtraAgainstExactObject for all extra props when subtyping against an exact object
Summary:
When we have `obj_with_random_extra_props ~> exact_obj`, we generate confusing prop-missing errors one by one. While this is technically correct as they are indeed missing in the extra object, this is very unhelpful and we tend to have tons of errors at a single location.

This was always a big problem, but it will get even worse and got hit more often once we fix the object literal unsoundness, so now it's the time to act.

This diff combines them into a single error, with a specialized error message including an explanation of "Exact objects do not accept extra props". I still keep the error `prop-missing` to make releases easier, but we can change it in the future as well.

Changelog: [errors] When an object with extra properties is passed to a place that expect an exact object, we will now generate a single error with all extra properties. The error message will list the extra properties, and state that "Exact objects do not accept extra props". In rare cases, the error locations might be moved.

Reviewed By: panagosg7

Differential Revision: D76372001

fbshipit-source-id: 8dedd4e2763f0a708863aca127f00fa74bf5bd91